Refactor #78: Split exercise_objectives.py into per-exercise modules
